move data for each test suite to it's own folder
