+++ tests/dummy-app/dummy-data/knexfile.ts
@@ -1,0 +1,13 @@
+// Update with your config settings.
+import { join } from "path";
+
+export default {
+  development: {
+    client: "sqlite3",
+    connection: {
+      filename: join(__dirname, "dev.sqlite3"),
+    },
+    useNullAsDefault: true,
+    debug: true,
+  },
+};
