+++ tests/test-suite/setup.ts
@@ -1,7 +1,7 @@
 import * as Knex from "knex";
 import app from "./test-app/app";
 import context from "./transaction";
-import knexfile from "../data/knexfile";
+import knexfile from "./test-data/knexfile";
 import Serializer from "./test-app/serializers/serializer";
 let knexInstance = Knex(knexfile[global["TEST_SUITE"]]);
 const createTransaction = (connection, callback): Promise<Knex.Transaction> => {
