+++ tests/test-suite/test-app/app.ts
@@ -7,7 +7,7 @@ import {
   UserManagementAddonOptions
 } from "./jsonapi-ts";
 
-import knexfile from "./../../data/knexfile";
+import knexfile from "./../test-data/knexfile";
 
 import login from "./callbacks/login";
