+++ tests/test-suite/test-data/camelCase/seeds/seed.ts
