+++ tests/test-suite/test-data/snake_case/seeds/seed.ts
